% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQOA.R
\name{readQOA}
\alias{readQOA}
\title{Read an QOA file}
\usage{
readQOA(qoa_path)
}
\arguments{
\item{qoa_path}{\link{character} (\strong{required}): Path to a stored qoa-file}
}
\value{
A list with the sample data, channels, samplerate and number of samples per channel
If the decoding went wrong the returned value is NULL.
}
\description{
Read an QOA file
}
\examples{
qoa_file <- system.file("extdata", "58_guitar_sarasate_stereo.qoa", package = "qoa")
qoa_data <- readQOA(qoa_file)
}
\author{
Johannes Friedrich
}
