% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoa-package.R
\docType{data}
\name{wav_example}
\alias{wav_example}
\title{raw values extracted from wav file}
\format{
[list] data as [matrix], samplerate, samples, channels
}
\description{
raw values extracted from wav file
}
\author{
Johannes Friedrich
}
