% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQOA.R
\name{writeQOA}
\alias{writeQOA}
\title{Write an QOA file}
\usage{
writeQOA(samples, samplerate, target = raw())
}
\arguments{
\item{samples}{\link{matrix} (\strong{required}): audio file represented by a integer matrix or array.}

\item{samplerate}{\link{integer} (\strong{required}): samplerate of the data given in argument 'samples'.}

\item{target}{\link{character} or \link{connections} or \link{raw}: Either name of the file to write, a binary connection or a raw vector (raw() - the default - is good enough) indicating that the output should be a raw vector.}
}
\value{
The result is either stored in a file (if target is a file name), in a raw vector (if target is a raw vector) or sent to a binary connection.
}
\description{
Write an QOA file
}
\examples{
## (1) Write to raw() -> see bytes
wav <- writeQOA(wav_example$data, wav_example$samplerate)
rawToChar(head(wav,4)) ##qoaf

\dontrun{
## (2) Write to a *.qoi file
writeQOA(wav_example$data, wav_example$samplerate, "wav_to_qoa.qoa")
}
}
\author{
Johannes Friedrich
}
